requirements.txt
pandas
numpy
matplotlib
seaborn
requests
transformers
torch
yfinance
python-dotenv
